Attached Meteor Methods on Server Side to Work With Semester's Functionality.
