+++ .meteor/versions
@@ -40,7 +40,6 @@ htmljs@1.0.11
 http@1.2.10
 hwillson:stub-collections@1.0.3
 id-map@1.0.9
-insecure@1.0.7
 jquery@1.11.10
 launch-screen@1.1.0
 livedata@1.0.18
