+++ server/main.ts
@@ -1,5 +1,6 @@
 import { Meteor } from 'meteor/meteor';
 
+import { ServerDBMethods } from './imports/DBMethods';
 import { loadCourses } from './imports/fixtures/courses';
 import { loadQueuedCourses } from './imports/fixtures/queuedcourses';
 import { loadSemesterCourses } from './imports/fixtures/semestercourses';
@@ -8,8 +9,9 @@ import { loadReqs } from './imports/fixtures/reqs';
 Starting up the universal database in the server
 */
 Meteor.startup(() => {
-  loadCourses();
-  loadQueuedCourses();
-  loadSemesterCourses();
-  loadReqs();
+  	loadCourses();
+  	loadQueuedCourses();
+  	loadSemesterCourses();
+  	loadReqs();
+  	return Meteor.methods(ServerDBMethods);
 });
